requirements.txt
alembic==1.13.2
amqp==5.2.0
aniso8601==9.0.1
attrs==23.2.0
bcrypt==4.1.3
billiard==3.6.4.0
black==22.3.0
blinker==1.8.2
celery==5.2.3
certifi==2024.6.2
charset-normalizer==2.0.12
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==6.3.2
Deprecated==1.2.14
flake8==4.0.1
Flask==2.3.3
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==20.1.0
idna==3.7
iniconfig==2.0.0
isort==5.10.1
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.3.7
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
mccabe==0.6.1
mypy-extensions==1.0.0
nulltype==2.3.1
packaging==24.1
pathspec==0.12.1
plaid-python==9.1.0
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.8.0
pytest==7.1.1
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
pytz==2024.1
PyYAML==6.0
redis==4.1.4
requests==2.27.1
six==1.16.0
SQLAlchemy==2.0.31
tomli==2.0.1
typing_extensions==4.12.2
urllib3==1.26.19
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
